Wrong type corrected and cosmetic fixes

+++ radio/src/telemetry/frsky.h
@@ -464,7 +464,7 @@ void telemetryInterrupt10ms(void);
 #if defined(CPUARM)
   typedef uint16_t frskyCellVoltage_t;
 #elif defined(FRSKY_HUB)
-  typedef uint16_t frskyCellVoltage_t;
+  typedef uint8_t frskyCellVoltage_t;
 #endif
 
 void frskySetCellVoltage(uint8_t battnumber, frskyCellVoltage_t cellVolts);
